requirements.txt
deepgram-sdk==4.7.0
pydub
google-generativeai
langchain-google-genai
python-dotenv
pandas
sendgrid
openpyxl
fastapi
uvicorn[standard]
gunicorn>=21.2.0 
python-multipart
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
PyJWT==2.8.0
supabase
cryptography
azure-storage-blob==12.24.0
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
azure-cosmos>=4.5.1
